backend/requirements.txt
Django==5.0.7
 djangorestframework==3.15.2
 djangorestframework-simplejwt==5.3.1
 python-decouple==3.8
 django-cors-headers==4.4.0
 Pillow==10.4.0
